Better actor linking
Changing the way things are constructed so that the listener explicitly
spawns a Session for each connection. The Session will in turn call
spawn_handler that creates an actor and (optionally) links it to the
session. If linked to the session, it will also be supervised and torn
down on session closure, both explicit closure and failures.