Even better actor linking and organization
The consumer of the Listener actor will how spawn independent Session
actors and can choose their linking as it see fit.

Also, separating the listener and session modules. The contents of the
tcp module could be renamed to stream and the listener/session modules
would be independent and only depend on stream.