Updating tcp (#32)
* fix: Typo

* Updating net actor with code from ractor_cluster.

The code based are obviously based on each other so it was fairly small
number of changes.

The FameReceiver and SessionAcceptor traits are gone, they're replaced
by the new Session messages and callback that has to be configured on
the Listener.

* Updating comments

* Consistent use of "frame" instead of "object". Mostly because that is
what was there from before.
* Dropping "node", it's for the ractor_cluster code.

* Better actor linking

Changing the way things are constructed so that the listener explicitly
spawns a Session for each connection. The Session will in turn call
spawn_handler that creates an actor and (optionally) links it to the
session. If linked to the session, it will also be supervised and torn
down on session closure, both explicit closure and failures.

* net/tcp example

* Better actor linking

This works more like advertised where the listener, session and
reader/writer halves are linked.

This leads to a tightly coupled system, more so that really necessary.

* Even better actor linking and organization

The consumer of the Listener actor will how spawn independent Session
actors and can choose their linking as it see fit.

Also, separating the listener and session modules. The contents of the
tcp module could be renamed to stream and the listener/session modules
would be independent and only depend on stream.

* Adding watchdog actor

This actor will monitor registered actors and kill/stop them after a
specified unless it receives a ping message.

* Watchdog fixes

* Updating after feedback

* Updating after feedback

* Adding docs on the watchdog module

* Rolling back some unnecessary changes

* Rolling back some unnecessary changes

Updating examples to new API

* ci: Enable on all branches

* Clippy fixes

* Fixing Clippy warning

* Comments and nits

* Adding missing async_trait

* cargo fmt --all

* Better docs

* Reducing tracing level

* Rust improvements

* Rust improvements

* Symbol re-export fixup

No semver breakages reported with `cargo semver-checks check-release`
with this change.

* watchdog: Making the ping call synchronous

* watchdog: Note on CPU starvation

* ci: Putting back branch specification